requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
bcrypt
pyjwt